RSS is turned into a dict
keys are 'title', 'description', 'image', 'published' and 'items'
'items' are a list of dicts with each entry having the keys of 'title', 'link',
'description' and 'thumbnail'. Each item represents an article.
The 'thumbnail' is the url of the thumbnail image for the article.

Added setup.py which makes this into a module for testing purposes.